requirements.txt
Django==4.2.20
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-filter==25.1
django-redis==5.4.0
PyJWT==2.10.1
Pillow==11.1.0
python-dotenv==1.1.0
uwsgi==2.0.28
celery==5.4.0
redis==5.2.1
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.0
cryptography==42.0.5